- About and contact is changed and merged.
- Rating system is added.
- Search function is complete(no AJAX yet).
- Database is updated, looks better.
-- All pages updated to work with new database.
- No-show law fields problem fixed.
- New footer merged.

+++ src/model/form_handlers/rating_handler.php
@@ -1,27 +1,36 @@
 <?php 
 
-	/*Get info*/
-	$rating_details = mysqli_query($con, "SELECT * FROM ratings WHERE lsp_id='$lsp_id'");
-	$ratings_info = mysqli_fetch_array($rating_details);
+/* Get count of reviews that the lawyer has gotten. */
+$rating_details = mysqli_query($con, "SELECT COUNT(*) FROM ratings WHERE lsp_id='$lsp_id'");
+$ratings_info = mysqli_fetch_array($rating_details);
+$reviews = $ratings_info[0];
 
-	$score = $ratings_info['score'];
-	$ratings = $ratings_info['numbers_of_ratings'];
-
-/* This will happen if the a user click the button to rate */
-if(isset($_POST['rate_button'])){ 
-	$rating = $_POST['rating']; 
-	$score += $rating;
-	$ratings++;
-	
-
-	$userid = $user['id'];
-	$query = mysqli_query($con, "DELETE FROM ratings WHERE lsp_id='$lsp_id'");
-	$query = mysqli_query($con, "INSERT INTO ratings VALUES ('$lsp_id', '$score', '$ratings')");
-	//$query = mysqli_query($con, "UPDATE ratings SET lsp_id='$lsp_id', score='$score', numbers_of_ratings='$ratings' WHERE lsp_id='$lsp_id'");
+/* Collects all reviews that the lawyer has. */
+$lspallreviews = mysqli_query($con, "SELECT * FROM ratings WHERE lsp_id='$lsp_id'");
 
 
+$sumofscore = mysqli_query($con, "SELECT SUM(score) FROM ratings WHERE lsp_id='$lsp_id'");
+$scoresum = mysqli_fetch_array($sumofscore);
 
+/* Gets the number that is used to decide if the user have posted a review before. */
+if (isset($user)) {
+	if ($user['usertype'] == 0) {
+		$user_id = $user['user_id'];	
+		$personal_reviews = mysqli_query($con, "SELECT COUNT(*) FROM ratings WHERE user_id='$user_id' AND lsp_id='$lsp_id'");
+		$userreviews = mysqli_fetch_array($personal_reviews);	
 
+		/* Reviews of lsp in all and if the user have reviewed the lsp. */
+		$userreview = $userreviews[0];
+
+		/* This will happen if the a user click the button to rate */
+		if(isset($_POST['rate_button'])){ 
+			$review = $_POST['review'];
+			$rating = $_POST['rating']; 
+
+			$query = mysqli_query($con, "DELETE FROM ratings WHERE lsp_id='$lsp_id' AND user_id='$user_id'");
+			$query = mysqli_query($con, "INSERT INTO ratings VALUES ('', '$lsp_id', '$user_id', '$review', '$rating')");
+		}
+	}
 }
 
 ?>
